python/requirements.txt
livekit-agents[openai,cartesia,deepgram,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
